+++ Sources/SwiftHTMLPreviewMacros/Plugin.swift
@@ -1,9 +1,0 @@
-import SwiftCompilerPlugin
-import SwiftSyntaxMacros
-
-@main
-struct SwiftHTMLPreviewMacrosPlugin: CompilerPlugin {
-    let providingMacros: [Macro.Type] = [
-        HTMLPreviewMacro.self,
-    ]
-}